HTMLBuilder

Description

All the data and type needed to the function

Documentation

creerPage :: String -> [Relation] -> [Infos] -> [(String, Style)] -> String

creerPage takes a subject and the graphs as parameters and return the html code for the page associated to the subject

type Relation = (String, String, String)

Relation links a subject and a target with a relation

type Infos = (String, Type, String)

Infos links a subject to some informations concerning the subject

data Style

Style defines the appereance for the HTML page Syntax:

Node nom [Parameter] [Style] -> représente une balise HTML. nom est le nom de la balise entre guillemets, [Parameter] est une liste d'attributs de la balise comme décrit plus bas, et Style est une liste des Noeuds qui seront dans la balise. Text texte -> insère un texte de votre choix Title -> insère le titre de la page IterateRelation [Style] -> permet de reproduire une fois par relation la série de balise contenue LinkName -> permet d'insérer le nom de la première relation active. Utilisable dans un noeud IterateRelation LinkTarget -> permet d'insérer la cible de la première relation active. Utilisable dans un noeud IterateRelation IterateInfos Type [Style] -> permet de reproduire une fois par élément du fichier info du type précisé la série de balise contenue Info -> permet d'insérer la cible de la première Info active. Utilisable dans un noeud IterateInfos

Constructors

Node String [Parameter] [Style]

Text String

Title

IterateRelation [Style]

LinkName

LinkTarget

IterateInfos Type [Style]

Info

data Type

The informations can be a Description or a Picture or an appereance for the HTML page

Constructors

Description

Image

HTMLType

data Parameter

represent a parameter in a HTML document use P for a string and L for a link

Constructors

Param String String

Link String

TitleP String

InfoP String