HTMLBuilderUtils

Description

create a module with the following function available

Documentation

getSujets :: [Relation] -> [String]

getSujets returns all the first and last element of a list of Relation with just one copy of each

shrinkToUniq :: [String] -> [String]

shrinkToUniq returns a list with just one copy of the elements of the list in parameter

getFirst :: Relation -> String

getFirst returns the first element of a Relation

getLast :: Relation -> String

getLast returns the last element of a Relation