Main

Description

create a Module Main

Documentation

main :: IO ()

the main function that read the graphs and create the site

site :: [Relation] -> [Infos] -> [(String, Style)] -> [String] -> Snap ()

select the part of the URL to load

echoHandler :: [Relation] -> [Infos] -> [(String, Style)] -> [String] -> Snap ()

filter a potential parse error in site

checkparam :: String -> [Relation] -> [Infos] -> [(String, Style)] -> [String] -> String

generate the page if it exit or a 404 error