Main

Documentation

f :: [Relation] -> String -> [Relation]

writer :: [Relation] -> [Infos] -> [(String, Style)] -> String -> IO ()

writer create a HTML page using the graphs and the subject of this page

main :: IO [()]

read all graphs, get the subjects and create the HTML Page linked to these subjects